fix(config): check each value of nested config properties during upsert
This patch fixes a bug (although not perfectly) in the `update`
subcommand. Since the config passed to `config:Upsert` *always* has
a non-nil `TLS` property (filled with default values, which means
`false` and empty strings), we need to, at the very least, check each
nested property to see if the values sent are non-empty.

Without this patch, changing *any* config option would wipe out the TLS
settings. See the regression test introduced in an earlier commit.

Note that this is still not quite perfect: it will still "reset" the
value of "disable host key verification" which isn't as problematic
since the default of `false` is the safer option.

A proper fix would likely involve changing the `Config` struct to use
pointers (so we can disambiguate between unset and empty) or an
alternate datatype to hold "updates" (such as a map), passed to the
`Upsert` function, or simply a slice of field names that were supplied
on the command line which are used to filter updates to the context.